flutter: restructured files
