+++ wetty-chat-flutter/lib/data/services/message_service.dart
@@ -1,0 +1,102 @@
+import 'dart:convert';
+import 'package:http/http.dart' as http;
+
+import '../../config/api_config.dart';
+import '../models/message_models.dart';
+
+/// Raw HTTP calls for message endpoints. No state.
+class MessageService {
+  Future<ListMessagesResponse> fetchMessages(
+    String chatId, {
+    int? max,
+    String? before,
+    String? after,
+  }) async {
+    final query = <String, String>{};
+    if (max != null) query['max'] = max.toString();
+    if (before != null && before.isNotEmpty) query['before'] = before;
+    if (after != null && after.isNotEmpty) query['after'] = after;
+    final uri = Uri.parse(
+      '$apiBaseUrl/chats/$chatId/messages',
+    ).replace(queryParameters: query.isEmpty ? null : query);
+    final response = await http.get(uri, headers: apiHeaders);
+    if (response.statusCode != 200) {
+      throw Exception(
+        'Failed to load messages: ${response.statusCode} ${response.body}',
+      );
+    }
+    return ListMessagesResponse.fromJson(
+      jsonDecode(response.body) as Map<String, dynamic>,
+    );
+  }
+
+  /// Fetches messages around [messageId] for deep linking.
+  Future<List<MessageItem>> fetchAround(
+    String chatId,
+    String messageId,
+  ) async {
+    final resBefore = await fetchMessages(chatId, max: 15, before: messageId);
+    final resAfter = await fetchMessages(chatId, max: 15, after: messageId);
+    return [...resBefore.messages, ...resAfter.messages];
+  }
+
+  Future<MessageItem> sendMessage(
+    String chatId,
+    String text, {
+    String? replyToId,
+  }) async {
+    final uri = Uri.parse('$apiBaseUrl/chats/$chatId/messages');
+    final clientGeneratedId =
+        '${DateTime.now().millisecondsSinceEpoch}-${Uri.base.hashCode}';
+    final body = <String, dynamic>{
+      'message': text,
+      'message_type': 'text',
+      'client_generated_id': clientGeneratedId,
+    };
+    if (replyToId != null) body['reply_to_id'] = int.parse(replyToId);
+    final response = await http.post(
+      uri,
+      headers: apiHeaders,
+      body: jsonEncode(body),
+    );
+    if (response.statusCode != 201) {
+      throw Exception(
+        'Failed to send message: ${response.statusCode} ${response.body}',
+      );
+    }
+    return MessageItem.fromJson(
+      jsonDecode(response.body) as Map<String, dynamic>,
+    );
+  }
+
+  Future<MessageItem> editMessage(
+    String chatId,
+    String messageId,
+    String newText,
+  ) async {
+    final uri = Uri.parse('$apiBaseUrl/chats/$chatId/messages/$messageId');
+    final response = await http.patch(
+      uri,
+      headers: apiHeaders,
+      body: jsonEncode({'message': newText}),
+    );
+    if (response.statusCode != 200) {
+      throw Exception(
+        'Failed to edit message: ${response.statusCode} ${response.body}',
+      );
+    }
+    return MessageItem.fromJson(
+      jsonDecode(response.body) as Map<String, dynamic>,
+    );
+  }
+
+  Future<void> deleteMessage(String chatId, String messageId) async {
+    final uri = Uri.parse('$apiBaseUrl/chats/$chatId/messages/$messageId');
+    final response = await http.delete(uri, headers: apiHeaders);
+    if (response.statusCode != 204) {
+      throw Exception(
+        'Failed to delete message: ${response.statusCode} ${response.body}',
+      );
+    }
+  }
+}
