+++ wetty-chat-flutter/lib/draft/chats.dart
@@ -1,0 +1,621 @@
+// import 'package:flutter/cupertino.dart';
+// import 'package:http/http.dart' as http;
+// import 'dart:convert';
+
+// import 'api_config.dart';
+// import 'draft_store.dart';
+// import 'models.dart';
+// import 'widgets.dart';
+
+// class ChatPage extends StatefulWidget {
+//   const ChatPage({super.key});
+
+//   @override
+//   State<ChatPage> createState() => _ChatPageState();
+// }
+
+// class _ChatPageState extends State<ChatPage> {
+//   String title = "Chats";
+//   List<ChatListItem> chats = [];
+//   bool isLoading = true;
+//   bool isLoadingMore = false;
+//   String? errorMessage;
+//   String? nextCursor;
+//   static const int _chatsSize = 11;
+//   late ScrollController _scrollController;
+//   late TextEditingController _nameController;
+
+//   bool get hasMoreChats => nextCursor != null && nextCursor!.isNotEmpty;
+
+//   @override
+//   void initState() {
+//     super.initState();
+//     _scrollController = ScrollController()..addListener(_onScroll);
+//     _nameController = TextEditingController();
+//     _loadChats();
+//   }
+
+//   @override
+//   void dispose() {
+//     _scrollController.removeListener(_onScroll);
+//     _scrollController.dispose();
+//     _nameController.dispose();
+//     super.dispose();
+//   }
+
+//   void _onScroll() {
+//     if (!hasMoreChats || isLoadingMore || isLoading) return;
+//     final pos = _scrollController.position;
+//     if (pos.pixels >= pos.maxScrollExtent - 200) {
+//       _loadMoreChats();
+//     }
+//   }
+
+//   // initial load
+//   Future<void> _loadChats() async {
+//     if (!mounted) return;
+//     setState(() {
+//       isLoading = true;
+//       errorMessage = null;
+//       nextCursor = null;
+//     });
+//     try {
+//       final res = await fetchChats(limit: _chatsSize);
+//       if (!mounted) return;
+//       setState(() {
+//         chats = res.chats;
+//         nextCursor = res.nextCursor;
+//         isLoading = false;
+//         errorMessage = null;
+//       });
+//     } catch (e) {
+//       if (!mounted) return;
+//       setState(() {
+//         isLoading = false;
+//         errorMessage = e.toString();
+//       });
+//     }
+//   }
+
+//   // check and load more chats when scrolling to the bottom of a page
+//   Future<void> _loadMoreChats() async {
+//     if (!hasMoreChats || isLoadingMore || chats.isEmpty) return;
+//     final lastId = chats.last.id;
+//     setState(() => isLoadingMore = true);
+//     try {
+//       final res = await fetchChats(limit: _chatsSize, after: lastId);
+//       if (!mounted) return;
+//       final existingIds = chats.map((c) => c.id).toSet();
+//       final newChats = res.chats
+//           .where((c) => !existingIds.contains(c.id))
+//           .toList();
+//       setState(() {
+//         chats = [...chats, ...newChats];
+//         nextCursor = res.nextCursor;
+//         isLoadingMore = false;
+//       });
+//     } catch (e) {
+//       if (!mounted) return;
+//       setState(() => isLoadingMore = false);
+//     }
+//   }
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return CupertinoPageScaffold(
+//       navigationBar: CupertinoNavigationBar(
+//         middle: Text(title),
+//         trailing: CupertinoButton(
+//           padding: EdgeInsets.zero,
+//           onPressed: addChat,
+//           child: const Icon(CupertinoIcons.square_pencil),
+//         ),
+//       ),
+//       child: SafeArea(child: _buildBody()),
+//     );
+//   }
+
+//   // body of chats page
+//   Widget _buildBody() {
+//     if (isLoading) {
+//       return const Center(child: CupertinoActivityIndicator());
+//     }
+//     if (errorMessage != null) {
+//       return Center(
+//         child: Padding(
+//           padding: const EdgeInsets.all(24.0),
+//           child: Column(
+//             mainAxisAlignment: MainAxisAlignment.center,
+//             children: [
+//               Text(errorMessage!, textAlign: TextAlign.center),
+//               const SizedBox(height: 16),
+//               CupertinoButton.filled(
+//                 onPressed: _loadChats,
+//                 child: const Text('Retry'),
+//               ),
+//             ],
+//           ),
+//         ),
+//       );
+//     }
+//     if (chats.isEmpty) {
+//       return const Center(child: Text('No chats yet'));
+//     }
+//     return ListView.builder(
+//       controller: _scrollController,
+//       itemCount: chats.length,
+//       itemBuilder: (context, index) {
+//         final chat = chats[index];
+//         final chatName = chat.name?.isNotEmpty == true
+//             ? chat.name!
+//             : 'Chat ${chat.id}';
+//         // Format date
+//         String? dateText;
+//         if (chat.lastMessageAt != null) {
+//           try {
+//             final dt = DateTime.parse(chat.lastMessageAt!);
+//             final now = DateTime.now();
+//             if (dt.day == now.day &&
+//                 dt.month == now.month &&
+//                 dt.year == now.year) {
+//               dateText =
+//                   '${dt.hour.toString().padLeft(2, '0')}:${dt.minute.toString().padLeft(2, '0')}';
+//             } else {
+//               dateText = '${dt.month}/${dt.day}';
+//             }
+//           } catch (_) {
+//             dateText = chat.lastMessageAt;
+//           }
+//         }
+//         final senderName = chat.lastMessage?.sender.name;
+//         final lastMsg = chat.lastMessage?.message;
+//         final unreadCount = chat.unreadCount;
+//         final hasMessage =
+//             (senderName != null && senderName.isNotEmpty) &&
+//             (lastMsg != null && lastMsg.isNotEmpty);
+
+//         return Column(
+//           mainAxisSize: MainAxisSize.min,
+//           children: [
+//             GestureDetector(
+//               onTap: () async {
+//                 await Navigator.push(
+//                   context,
+//                   CupertinoPageRoute(
+//                     builder: (_) => ChatDetailPage(
+//                       chatId: chat.id,
+//                       chatName: chat.name ?? 'Chat ${chat.id}',
+//                     ),
+//                   ),
+//                 );
+//                 // Refresh so draft indicators update immediately
+//                 if (mounted) setState(() {});
+//               },
+//               behavior: HitTestBehavior.opaque,
+//               child: Padding(
+//                 padding: const EdgeInsets.symmetric(
+//                   horizontal: 16,
+//                   vertical: 8,
+//                 ),
+//                 // each chat item
+//                 child: Row(
+//                   children: [
+//                     // TODO: change avatar to group avatar
+//                     // Avatar
+//                     Container(
+//                       width: 52,
+//                       height: 52,
+//                       decoration: BoxDecoration(
+//                         color: CupertinoColors.systemGrey4,
+//                         shape: BoxShape.circle,
+//                       ),
+//                       alignment: Alignment.center,
+//                       child: Text(
+//                         chatName.isNotEmpty ? chatName[0].toUpperCase() : '?',
+//                         style: const TextStyle(
+//                           fontSize: 18,
+//                           fontWeight: FontWeight.w600,
+//                           color: CupertinoColors.white,
+//                         ),
+//                       ),
+//                     ),
+//                     const SizedBox(width: 12),
+//                     // Chat info
+//                     Expanded(
+//                       child: Column(
+//                         crossAxisAlignment: CrossAxisAlignment.start,
+//                         children: [
+//                           // Top row: chat name + date
+//                           Row(
+//                             children: [
+//                               Expanded(
+//                                 child: Text(
+//                                   chatName,
+//                                   style: const TextStyle(
+//                                     fontSize: 16,
+//                                     fontWeight: FontWeight.w600,
+//                                   ),
+//                                   maxLines: 1,
+//                                   overflow: TextOverflow.ellipsis,
+//                                 ),
+//                               ),
+//                               if (dateText != null)
+//                                 Padding(
+//                                   padding: const EdgeInsets.only(left: 8),
+//                                   child: Text(
+//                                     dateText,
+//                                     style: TextStyle(
+//                                       fontSize: 12,
+//                                       color: CupertinoColors.secondaryLabel
+//                                           .resolveFrom(context),
+//                                     ),
+//                                   ),
+//                                 ),
+//                             ],
+//                           ),
+//                           const SizedBox(height: 3),
+//                           // Bottom row: draft or sender: last message + unread count
+//                           Builder(
+//                             builder: (context) {
+//                               final draft = DraftStore.instance.getDraft(
+//                                 chat.id,
+//                               );
+//                               if (draft != null) {
+//                                 return Row(
+//                                   children: [
+//                                     Expanded(
+//                                       child: Text.rich(
+//                                         TextSpan(
+//                                           children: [
+//                                             const TextSpan(
+//                                               text: '[Draft] ',
+//                                               style: TextStyle(
+//                                                 fontSize: 13,
+//                                                 color: CupertinoColors
+//                                                     .destructiveRed,
+//                                                 fontWeight: FontWeight.w500,
+//                                               ),
+//                                             ),
+//                                             TextSpan(
+//                                               text: draft,
+//                                               style: TextStyle(
+//                                                 fontSize: 13,
+//                                                 color: CupertinoColors
+//                                                     .secondaryLabel
+//                                                     .resolveFrom(context),
+//                                               ),
+//                                             ),
+//                                           ],
+//                                         ),
+//                                         maxLines: 1,
+//                                         overflow: TextOverflow.ellipsis,
+//                                       ),
+//                                     ),
+//                                     if (unreadCount > 0)
+//                                       Container(
+//                                         margin: const EdgeInsets.only(left: 8),
+//                                         padding: const EdgeInsets.symmetric(
+//                                           horizontal: 6,
+//                                           vertical: 2,
+//                                         ),
+//                                         decoration: BoxDecoration(
+//                                           color: CupertinoColors.activeBlue,
+//                                           borderRadius: BorderRadius.circular(
+//                                             10,
+//                                           ),
+//                                         ),
+//                                         child: Text(
+//                                           '$unreadCount',
+//                                           style: const TextStyle(
+//                                             color: CupertinoColors.white,
+//                                             fontSize: 11,
+//                                             fontWeight: FontWeight.w600,
+//                                           ),
+//                                         ),
+//                                       ),
+//                                   ],
+//                                 );
+//                               }
+//                               return Row(
+//                                 children: [
+//                                   Expanded(
+//                                     child: hasMessage
+//                                         ? Text.rich(
+//                                             TextSpan(
+//                                               children: [
+//                                                 TextSpan(
+//                                                   text: '$senderName: ',
+//                                                   style: const TextStyle(
+//                                                     fontWeight: FontWeight.bold,
+//                                                   ),
+//                                                 ),
+//                                                 TextSpan(text: lastMsg),
+//                                               ],
+//                                             ),
+//                                             maxLines: 1,
+//                                             overflow: TextOverflow.ellipsis,
+//                                             style: TextStyle(
+//                                               fontSize: 13,
+//                                               color: CupertinoColors
+//                                                   .secondaryLabel
+//                                                   .resolveFrom(context),
+//                                             ),
+//                                           )
+//                                         : Text(
+//                                             'No messages yet',
+//                                             maxLines: 1,
+//                                             overflow: TextOverflow.ellipsis,
+//                                             style: TextStyle(
+//                                               fontSize: 13,
+//                                               color: CupertinoColors
+//                                                   .secondaryLabel
+//                                                   .resolveFrom(context),
+//                                             ),
+//                                           ),
+//                                   ),
+//                                   if (unreadCount > 0)
+//                                     Container(
+//                                       margin: const EdgeInsets.only(left: 8),
+//                                       padding: const EdgeInsets.symmetric(
+//                                         horizontal: 6,
+//                                         vertical: 2,
+//                                       ),
+//                                       decoration: BoxDecoration(
+//                                         color: CupertinoColors.activeBlue,
+//                                         borderRadius: BorderRadius.circular(10),
+//                                       ),
+//                                       child: Text(
+//                                         '$unreadCount',
+//                                         style: const TextStyle(
+//                                           color: CupertinoColors.white,
+//                                           fontSize: 11,
+//                                           fontWeight: FontWeight.w600,
+//                                         ),
+//                                       ),
+//                                     ),
+//                                 ],
+//                               );
+//                             },
+//                           ),
+//                         ],
+//                       ),
+//                     ),
+//                     const SizedBox(width: 8),
+//                     // Disclosure indicator
+//                     const Icon(
+//                       CupertinoIcons.chevron_right,
+//                       size: 16,
+//                       color: CupertinoColors.systemGrey3,
+//                     ),
+//                   ],
+//                 ),
+//               ),
+//             ),
+//             // Bottom separator for every item (including the last)
+//             Padding(
+//               padding: const EdgeInsets.only(left: 72),
+//               child: Divider(height: 0.5, color: CupertinoColors.separator),
+//             ),
+//           ],
+//         );
+//       },
+//     );
+//   }
+
+//   // get chats
+//   Future<ListChatsResponse> fetchChats({int? limit, String? after}) async {
+//     final query = <String, String>{};
+//     if (limit != null) query['limit'] = limit.toString();
+//     if (after != null && after.isNotEmpty) query['after'] = after;
+//     final uri = Uri.parse(
+//       '$apiBaseUrl/chats',
+//     ).replace(queryParameters: query.isEmpty ? null : query);
+//     final response = await http.get(uri, headers: apiHeaders);
+//     if (response.statusCode != 200) {
+//       throw Exception(
+//         'Failed to load chats: ${response.statusCode} ${response.body}',
+//       );
+//     }
+//     return ListChatsResponse.fromJson(
+//       jsonDecode(response.body) as Map<String, dynamic>,
+//     );
+//   }
+
+//   // create chat
+//   Future<http.Response> createChat({String? name}) async {
+//     final url = Uri.parse('$apiBaseUrl/group');
+//     return http.post(
+//       url,
+//       headers: apiHeaders,
+//       body: jsonEncode({"name": name}),
+//     );
+//   }
+
+//   Future<void> addChat() async {
+//     final newChat = await Navigator.push<ChatListItem>(
+//       context,
+//       CupertinoPageRoute(builder: (_) => NewChatPage(createChat: createChat)),
+//     );
+//     if (newChat != null && mounted) {
+//       setState(() => chats.insert(0, newChat));
+//       _showToast('Chat created');
+//     }
+//   }
+
+//   /// Shows a brief toast-style overlay since Cupertino has no SnackBar.
+//   void _showToast(String message) {
+//     final overlay = Navigator.of(context).overlay;
+//     if (overlay == null) return;
+//     late OverlayEntry entry;
+//     entry = OverlayEntry(
+//       builder: (_) => Positioned(
+//         bottom: 80,
+//         left: 24,
+//         right: 24,
+//         child: _ToastWidget(message: message, onDismiss: () => entry.remove()),
+//       ),
+//     );
+//     overlay.insert(entry);
+//   }
+// }
+
+// // New Chat page
+// class NewChatPage extends StatefulWidget {
+//   const NewChatPage({super.key, required this.createChat});
+//   final Future<http.Response> Function({String? name}) createChat;
+
+//   @override
+//   State<NewChatPage> createState() => _NewChatPageState();
+// }
+
+// class _NewChatPageState extends State<NewChatPage> {
+//   final _nameController = TextEditingController();
+//   bool _isCreating = false;
+
+//   @override
+//   void dispose() {
+//     _nameController.dispose();
+//     super.dispose();
+//   }
+
+//   Future<void> _onCreate() async {
+//     if (_isCreating) return;
+//     setState(() => _isCreating = true);
+//     final name = _nameController.text.trim();
+//     try {
+//       final response = await widget.createChat(
+//         name: name.isEmpty ? null : name,
+//       );
+//       if (!mounted) return;
+//       if (response.statusCode == 201) {
+//         final body = jsonDecode(response.body) as Map<String, dynamic>;
+//         final id = body['id']?.toString() ?? '';
+//         final createdName = body['name'] as String?;
+//         Navigator.of(context).pop(ChatListItem(id: id, name: createdName));
+//       } else {
+//         setState(() => _isCreating = false);
+//         _showError('Server error: ${response.body}');
+//       }
+//     } catch (e) {
+//       if (!mounted) return;
+//       setState(() => _isCreating = false);
+//       _showError('Network error: $e');
+//     }
+//   }
+
+//   void _showError(String message) {
+//     showCupertinoDialog(
+//       context: context,
+//       builder: (ctx) => CupertinoAlertDialog(
+//         title: const Text('Error'),
+//         content: Text(message),
+//         actions: [
+//           CupertinoDialogAction(
+//             onPressed: () => Navigator.of(ctx).pop(),
+//             child: const Text('OK'),
+//           ),
+//         ],
+//       ),
+//     );
+//   }
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return CupertinoPageScaffold(
+//       navigationBar: CupertinoNavigationBar(
+//         middle: const Text('New Chat'),
+//         leading: CupertinoButton(
+//           padding: EdgeInsets.zero,
+//           onPressed: () => Navigator.of(context).pop(),
+//           child: const Text('Back'),
+//         ),
+//       ),
+//       child: GestureDetector(
+//         onTap: () => FocusScope.of(context).unfocus(),
+//         child: SafeArea(
+//           child: Padding(
+//             padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 24),
+//             child: Column(
+//               crossAxisAlignment: CrossAxisAlignment.start,
+//               children: [
+//                 const Text(
+//                   'Chat Name',
+//                   style: TextStyle(fontSize: 15, fontWeight: FontWeight.w600),
+//                 ),
+//                 const SizedBox(height: 8),
+//                 CupertinoTextField(
+//                   controller: _nameController,
+//                   placeholder: 'Optional',
+//                   padding: const EdgeInsets.symmetric(
+//                     horizontal: 12,
+//                     vertical: 12,
+//                   ),
+//                 ),
+//                 const SizedBox(height: 24),
+//                 SizedBox(
+//                   width: double.infinity,
+//                   child: CupertinoButton.filled(
+//                     onPressed: _isCreating ? null : _onCreate,
+//                     child: _isCreating
+//                         ? const CupertinoActivityIndicator(
+//                             color: CupertinoColors.white,
+//                           )
+//                         : const Text('Create'),
+//                   ),
+//                 ),
+//               ],
+//             ),
+//           ),
+//         ),
+//       ),
+//     );
+//   }
+// }
+
+// /// Simple animated toast widget for Cupertino context.
+// class _ToastWidget extends StatefulWidget {
+//   const _ToastWidget({required this.message, required this.onDismiss});
+//   final String message;
+//   final VoidCallback onDismiss;
+
+//   @override
+//   State<_ToastWidget> createState() => _ToastWidgetState();
+// }
+
+// class _ToastWidgetState extends State<_ToastWidget> {
+//   double _opacity = 0;
+
+//   @override
+//   void initState() {
+//     super.initState();
+//     Future.microtask(() {
+//       if (mounted) setState(() => _opacity = 1);
+//     });
+//     Future.delayed(const Duration(seconds: 2), () {
+//       if (mounted) setState(() => _opacity = 0);
+//       Future.delayed(const Duration(milliseconds: 300), widget.onDismiss);
+//     });
+//   }
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return IgnorePointer(
+//       child: AnimatedOpacity(
+//         opacity: _opacity,
+//         duration: const Duration(milliseconds: 300),
+//         child: Container(
+//           padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+//           decoration: BoxDecoration(
+//             color: CupertinoColors.systemGrey.withAlpha(230),
+//             borderRadius: BorderRadius.circular(10),
+//           ),
+//           child: Text(
+//             widget.message,
+//             textAlign: TextAlign.center,
+//             style: const TextStyle(color: CupertinoColors.white, fontSize: 14),
+//           ),
+//         ),
+//       ),
+//     );
+//   }
+// }
