+++ wetty-chat-flutter/lib/draft/main.dart
@@ -1,0 +1,24 @@
+// import 'package:flutter/cupertino.dart';
+
+// import 'chats.dart';
+// import 'draft_store.dart';
+
+// void main() async {
+//   WidgetsFlutterBinding.ensureInitialized();
+//   await DraftStore.instance.init();
+//   runApp(const MyApp());
+// }
+
+// class MyApp extends StatelessWidget {
+//   const MyApp({super.key});
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return const CupertinoApp(
+//       theme: CupertinoThemeData(
+//         brightness: Brightness.light,
+//       ),
+//       home: ChatPage(),
+//     );
+//   }
+// }
